+++ src/ConnectionFactory/RedisConnectionFactory.php
@@ -6,6 +6,12 @@ use Redis;
 
 class RedisConnectionFactory implements ConnectionFactory
 {
+    private int $optSerializer = Redis::SERIALIZER_NONE;
+
+    public function __construct(int $optSerializer = Redis::SERIALIZER_NONE)
+    {
+        $this->optSerializer = $optSerializer;
+    }
     /**
      * @return Redis
      */
@@ -13,6 +19,8 @@ class RedisConnectionFactory implements ConnectionFactory
     {
         $redis = new Redis();
         $redis->connect($host, $port, $timeout);
+        $redis->setOption(Redis::OPT_SERIALIZER, $this->optSerializer);
+
         return $redis;
     }
 }